requirements.txt
python-decouple==3.7
selenium==4.8.0